Add simtime, position. Return a boolean on loadWorld.

+++ src/TestWrapper.cpp
@@ -25,10 +25,35 @@ TestWrapper::TestWrapper() :
 	this->terminateGZClient = false;
 }
 
+/**
+ * START RPC METHOD IMPLEMENTATIONS
+ */
+
 //////////////////////////////////////////////////
 bool TestWrapper::onObject(const std::string& object, const std::string& surface)
 {
 	return OnEntity(this->world->GetModel(object), this->world->GetModel(surface));
+}
+
+//////////////////////////////////////////////////
+Json::Value TestWrapper::getPosition(const std::string& object)
+{
+	double position[] =
+	{ 0, 0, 0 };
+	physics::ModelPtr model = this->world->GetModel(object);
+	if (model != NULL)
+	{
+		position[0] = model->GetBoundingBox().GetCenter().x;
+		position[1] = model->GetBoundingBox().GetCenter().y;
+		position[2] = model->GetBoundingBox().GetCenter().z;
+	}
+	return Json::Value(position);
+}
+
+//////////////////////////////////////////////////
+double TestWrapper::getSimtime()
+{
+	return this->world->GetSimTime().Double();
 }
 
 //////////////////////////////////////////////////
@@ -38,7 +63,7 @@ void TestWrapper::resetWorld()
 }
 
 //////////////////////////////////////////////////
-void TestWrapper::loadWorld(const std::string& worldPath)
+bool TestWrapper::loadWorld(const std::string& worldPath)
 {
 	int waitCount = 0;
 	int maxWaitCount = 20000;
@@ -52,57 +77,33 @@ void TestWrapper::loadWorld(const std::string& worldPath)
 	if (!sdf::init(sdf))
 	{
 		gzerr << "Unable to initialize sdf\n";
-		return;
+		return false;
 	}
 
 	if (!sdf::readFile(worldPath, sdf))
 	// if (!sdf::readFile("/usr/share/gazebo-1.9/worlds/empty.world", sdf))
 	{
 		gzerr << "Unable to read SDF file [" << worldPath << "]\n";
-		return;
+		return false;
 	}
 	sdf::ElementPtr worldElem = sdf->root->GetElement("world");
 	if (worldElem)
 	{
-		// Get the currently running world
-		gazebo::physics::WorldPtr old_default_world = gazebo::physics::get_world();
-		// Push an empty world to the g_worlds list
-		gazebo::physics::WorldPtr new_world = gazebo::physics::create_world("default");
-		if (old_default_world)
-		{
-			gzmsg << "Got an old default world. Stopping world ...\n";
-			old_default_world->Stop();
-			// Get access to gazebo::g_worlds, which is defined in PhysicsIface.cc
-			g_worlds.erase(g_worlds.begin());
-			// 1 World should be left ...
-			gzmsg << "Running worlds " << g_worlds.size() << "\n";
-		}
-		// Create the world
-		try
-		{
-			gzmsg << "Loading world\n";
-			gazebo::physics::load_world(new_world, worldElem);
-			gzmsg << "Init world\n";
-			gazebo::physics::init_world(new_world);
-			gzmsg << "Run world\n";
-			gazebo::physics::run_world(new_world);
-			this->world = new_world;
-			killGazeboGUI();
-		} catch (gazebo::common::Exception &e)
-		{
-			gzthrow("Failed to load the World\n" << e);
-		}
+		gazebo::physics::WorldPtr new_world = replaceWorld();
+		bootWorld(new_world, worldElem);
+		return true;
 	}
-	else
-	{
-		gzerr << "World SDF Element can't be loaded\n";
-	}
+	gzerr << "World SDF Element can't be loaded\n";
+	return false;
 }
+
+/**
+ * START GAZEBO TESTING IMPLEMENTATIONS
+ */
 
 //////////////////////////////////////////////////
 bool TestWrapper::OnEntity(physics::EntityPtr entity, physics::EntityPtr onEntity)
 {
-
 	// Fail on null
 	if (entity == NULL || onEntity == NULL)
 	{
@@ -125,6 +126,49 @@ bool TestWrapper::OnEntity(physics::EntityPtr entity, physics::EntityPtr onEntit
 			&& onEntityBox.GetCenter().y - onEntityBox.GetYLength() / 2 <= center.y
 			&& center.z - entityBox.GetZLength() / 2 - onEntityBox.GetCenter().z - onEntityBox.GetZLength() / 2
 					<= ON_ENTITY_TOLERANCE;
+}
+
+/**
+ * START GAZEBO LIFECYCLE CONTROL
+ */
+
+//////////////////////////////////////////////////
+gazebo::physics::WorldPtr TestWrapper::replaceWorld()
+{
+	// Get the currently running world
+	gazebo::physics::WorldPtr old_default_world = gazebo::physics::get_world();
+	// Push an empty world to the g_worlds list
+	gazebo::physics::WorldPtr new_world = gazebo::physics::create_world("default");
+	if (old_default_world)
+	{
+		gzmsg << "Got an old default world. Stopping world ...\n";
+		old_default_world->Stop();
+		// Get access to gazebo::g_worlds, which is defined in PhysicsIface.cc
+		g_worlds.erase(g_worlds.begin());
+		// 1 World should be left ...
+		gzmsg << "Running worlds " << g_worlds.size() << "\n";
+	}
+	return new_world;
+}
+
+//////////////////////////////////////////////////
+void TestWrapper::bootWorld(const gazebo::physics::WorldPtr& new_world, const sdf::ElementPtr& worldElem)
+{
+	// Create the world
+	try
+	{
+		gzmsg << "Loading world\n";
+		gazebo::physics::load_world(new_world, worldElem);
+		gzmsg << "Init world\n";
+		gazebo::physics::init_world(new_world);
+		gzmsg << "Run world\n";
+		gazebo::physics::run_world(new_world);
+		this->world = new_world;
+		killGazeboGUI();
+	} catch (gazebo::common::Exception &e)
+	{
+		gzthrow("Failed to load the World\n" << e);
+	}
 }
 
 //////////////////////////////////////////////////
@@ -160,15 +204,20 @@ void TestWrapper::startGazeboGUI(char **_argv)
 	}
 }
 
+//////////////////////////////////////////////////
 void TestWrapper::stopGazeboGUI()
 {
 	this->terminateGZClient = true;
 	killGazeboGUI();
 }
 
+//////////////////////////////////////////////////
 void TestWrapper::killGazeboGUI()
 {
-	kill(this->gzclientPID, SIGKILL);
+	if (this->gzclientPID)
+	{
+		kill(this->gzclientPID, SIGKILL);
+	}
 }
 
 } /* namespace gazebo */
@@ -191,9 +240,12 @@ int main(int _argc, char **_argv)
 	for (int n = 0; n < _argc; n++)
 	{
 		std::string arg(_argv[n]);
-		if(arg == "-hl") {
+		if (arg == "-hl")
+		{
 			headless = true;
-		} else {
+		}
+		else
+		{
 			args.push_back(_argv[n]);
 		}
 	}
